create_vector_db/requirements.txt
google-generativeai
langchain
langchain-chroma
langchain-community
langchain-core
langchain-google-genai
langchain-experimental
langchain-text-splitters
langchain-experimental